backend/requirements.txt
asposestorage==1.0.2
Flask==3.0.2
Flask_Cors==4.0.0
oracledb==2.1.0
python-dotenv==1.0.1
Requests==2.31.0
SQLAlchemy==2.0.28
